Change 'Display mode' label
While testing the math plugin it was apparent that the current label is not
descriptive enough. Simply enabling or disabling the 'Display mode' did not
suggest the users that it would actually display the math formula on its own
line with proper spacing, or inline with the whole text. This commit aims to
improve the user experience by changing the title to more descriptive one

Trello card:
https://trello.com/c/y3tFZQDE